Migration管理できるように修正
