+++ rest_base_sample/package.json
@@ -32,6 +32,7 @@
     "@nestjs/typeorm": "^10.0.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
+    "dotenv": "^16.3.1",
     "firebase-admin": "^11.10.1",
     "hbs": "^4.2.0",
     "pg": "^8.11.2",
