+++ rest_base_sample/src/restaurants/restaurants.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
-import { Restaurant } from './restaurants.entity'
+import { Restaurant } from './entities/restaurant.entity'
 
 @Injectable()
 export class RestaurantsService {
