+++ rest_base_sample/src/seats/dto/create-seat.dto.ts
@@ -1,0 +1,17 @@
+import { IsNumber, IsNotEmpty } from 'class-validator'
+
+export class CreateSeatDto {
+  @IsNotEmpty()
+  @IsNumber()
+  number_of_seats: number
+
+  @IsNotEmpty()
+  start_at: string
+
+  @IsNotEmpty()
+  end_at: string
+
+  @IsNotEmpty()
+  @IsNumber()
+  restaurant_id: number
+}
