+++ rest_base_sample/src/seats/seats.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { SeatsController } from './seats.controller'
 import { SeatsService } from './seats.service'
-import { Seat } from './seats.entity'
+import { Seat } from './entities/seat.entity'
 
 @Module({
   imports: [TypeOrmModule.forFeature([Seat])],
